Disambiguate types
