+++ cedar-policy-validator/src/custom_schema.rs
@@ -2,3 +2,4 @@ mod ast;
 mod err;
 pub mod parser;
 mod test;
+pub mod to_json_schema;
